Mostly reverts, since a branch had to be merged that contained hasty changes that weren't well discussed.
Fixes the map, as it had a few bugs after the previous edits.
Captain once again receives access via proc.
Captain, HoS, and CE exclusive accesses were removed.
RD regained access to AI upload, as that is a large part of it's role.
Dual authentication doors removed. (If we really want this, there are cleaner ways)